Converted Friend Classes to Getters / Setters
One of the things I learned recently was that friend classes should generally be reserved if only one class needs to access another class's protected members. Getters and setters are a better method for encapsulation, so I went and removed as many friend classes as I could and replaced them with getter and setter functions.